Change archive dir

+++ bin/prepare-archive.sh
@@ -9,10 +9,18 @@ fi
 
 ARCHIVE_LOCATION="$1"
 
+cd "$ARCHIVE_DIR"
+
+if [[ -e .stacklift.ready ]] ; then
+    exit 0
+fi
+
 if [[ "$ARCHIVE_LOCATION" == s3://* ]] ; then
   aws s3 cp "$ARCHIVE_LOCATION" - | tar xz
 else
-  cp -pr "$ARCHIVE_LOCATION/." .
+  tar xzf "$ARCHIVE_LOCATION"
 fi
 
 pipenv sync --dev
+
+touch .stacklift.ready
